Exception handling - delete

+++ src/main/java/com/java_course/webservicesspringbootjpahibernate/services/UserService.java
@@ -4,11 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.java_course.webservicesspringbootjpahibernate.entities.User;
 
 import com.java_course.webservicesspringbootjpahibernate.repositories.UserRepository;
+import com.java_course.webservicesspringbootjpahibernate.services.exceptions.DatabaseException;
 import com.java_course.webservicesspringbootjpahibernate.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,15 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id))
+                throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+        } catch (ResourceNotFoundException e) {
+            throw e;
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
